+++ taxel-gui/src/main.rs
@@ -4,6 +4,7 @@ mod widgets;
 use crate::app::TaxelApp;
 use eframe::egui::ViewportBuilder;
 use log::debug;
+
 fn main() -> Result<(), anyhow::Error> {
     // Use hot reload in debug mode
     #[cfg(debug_assertions)]